m3u8-downloader/pyproject.toml
[tool.poetry]
name = "m3u8-downloader"
version = "0.1.0"
description = ""
authors = ["happyhappy-jun <bjyoon513@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^5.2.1"
fire = "^0.6.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.2"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
